requirements.txt
appnope==0.1.0
asn1crypto==0.22.0
attrs==15.2.0
Automat==0.6.0
baidu-aip==2.2.2.0
beautifulsoup4==4.6.0
bleach==2.1.2
bs4==0.0.1
# certifi==2017.11.5
cffi==1.10.0
chardet==3.0.4
colorama==0.3.9
constantly==15.1.0
cryptography==1.8.1
cssselect==1.0.1
cycler==0.10.0
decorator==4.2.1
demjson==2.2.4
entrypoints==0.2.3
et-xmlfile==1.0.1
fake-useragent==0.1.10
html5lib==1.0.1
hyperlink==17.1.1
idna==2.6
incremental==16.10.1
ipykernel==4.8.2
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.2
jdcal==1.4
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-core==4.4.0
lxml==3.8.0
MarkupSafe==1.0
matplotlib==2.1.2
mistune==0.8.3
Naked==0.1.31
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.4.0
numexpr==2.6.4
numpy==1.14.0
opencv-python==3.4.0.12
openpyxl==2.5.3
packaging==16.8
pandas==0.22.0
pandocfilters==1.4.2
parse==1.8.2
parsel==1.2.0
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
Pillow==5.0.0
pinyin==0.4.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pyasn1==0.2.3
pyasn1-modules==0.0.8
pycparser==2.18
# pycrypto==2.6.1
PyDispatcher==2.0.5
pyee==5.0.0
Pygments==2.2.0
pymongo==3.6.0
PyMySQL==0.8.0
pyOpenSSL==17.0.0
pyparsing==2.2.0
pypinyin==0.30.0
pyppeteer==0.0.17
PyQt5==5.10.1
pyquery==1.4.0
PySocks==1.6.8
pytesseract==0.1.8
python-dateutil==2.6.1
pytz==2018.3
PyYAML==3.12
pyzmq==17.0.0
QScintilla==2.10.4
qtconsole==4.3.1
queuelib==1.4.2
redis==2.10.6
requests==2.18.4
requests-html==0.9.0
Scrapy==1.5.0
scrapy-redis==0.6.8
scrapyd==1.2.0
scrapyd-client==1.1.0
selenium==2.42.1
Send2Trash==1.5.0
service-identity==17.0.0
shellescape==3.4.1
simplegeneric==0.8.1
sip==4.19.8
six==1.10.0
tables==3.4.2
terminado==0.8.1
tesseract==0.1.3
testpath==0.3.1
tornado==4.5.3
traitlets==4.3.2
# tushare==1.1.5
Twisted==17.5.0
urllib3==1.22
w3lib==1.17.0
wcwidth==0.1.7
webencodings==0.5.1
websockets==4.0.1
widgetsnbextension==3.1.4
wincertstore==0.2
xlrd==1.1.0
XlsxWriter==1.0.4
xlwt==1.3.0
zope.interface==4.4.2
